*moved dao runtime config to dao_config module
